+++ src/components/atoms/Button/Button.stories.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { withKnobs } from '@storybook/addon-knobs';
 
-import Button from '.';
+import Button from './Button';
 
 export default {
     title: 'Atoms/Button',
