+++ src/components/atoms/Button/Button.test.js
@@ -1,4 +1,4 @@
-import Button from '.';
+import Button from './Button';
 import runCommonTests from '../../../../test/commonTests';
 
 runCommonTests(Button);
