+++ src/components/atoms/Button/index.js
@@ -1,41 +1,3 @@
-import React from 'react';
-import {
-    func,
-    oneOf,
-    string,
-} from 'prop-types';
-
-import { StyledButton } from '../../styled-elements';
-
-const Button = (props) => {
-    const {
-        onClick,
-        text,
-        type,
-    } = props;
-
-    const buttonText = text || (type === 'submit' && 'Submit');
-
-    return (
-        <StyledButton
-            onClick={onClick}
-            type={type}
-        >
-            {buttonText}
-        </StyledButton>
-    );
-};
-
-Button.defaultProps = {
-    onClick: () => null,
-    text: '',
-    type: 'button',
-};
-
-Button.propTypes = {
-    onClick: func,
-    text: string,
-    type: oneOf([ 'button', 'submit' ]),
-};
+import Button from './Button';
 
 export default Button;
