+++ src/components/atoms/Card/Card.test.js
@@ -1,4 +1,4 @@
-import Card from '.';
+import Card from './Card';
 import runCommonTests from '../../../../test/commonTests';
 
 runCommonTests(Card);
