+++ src/components/atoms/Overlay/Overlay.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import {
+    func,
+} from 'prop-types';
+
+import {
+    StyledOverlay,
+} from '../../styled-elements';
+
+const Overlay = ({ onClick }) => (
+    <StyledOverlay onClick={onClick}></StyledOverlay>
+);
+
+Overlay.defaultProps = {
+    onClick: () => null,
+};
+
+Overlay.propTypes = {
+    onClick: func,
+};
+
+export default Overlay;
