+++ src/components/atoms/Overlay/Overlay.test.js
@@ -1,4 +1,4 @@
-import ScreenOverlay from '.';
+import ScreenOverlay from './Overlay';
 import runCommonTests from '../../../../test/commonTests';
 
 runCommonTests(ScreenOverlay);
