+++ src/components/atoms/Overlay/index.js
@@ -1,22 +1,3 @@
-import React from 'react';
-import {
-    func,
-} from 'prop-types';
-
-import {
-    StyledOverlay,
-} from '../../styled-elements';
-
-const Overlay = ({ onClick }) => (
-    <StyledOverlay onClick={onClick}></StyledOverlay>
-);
-
-Overlay.defaultProps = {
-    onClick: () => null,
-};
-
-Overlay.propTypes = {
-    onClick: func,
-};
+import Overlay from './Overlay';
 
 export default Overlay;
