+++ src/components/atoms/Panel/Panel.test.js
@@ -1,4 +1,4 @@
-import Panel from '.';
+import Panel from './Panel';
 import runCommonTests from '../../../../test/commonTests';
 
 runCommonTests(Panel);
