+++ src/components/atoms/Text/Text.test.js
@@ -1,4 +1,4 @@
-import Text from '.';
+import Text from './Text';
 import runCommonTests from '../../../../test/commonTests';
 
 runCommonTests(Text);
