+++ src/components/molecules/Drawer/Drawer.stories.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import { withKnobs } from '@storybook/addon-knobs';
 
-import Drawer from '.';
+import Drawer from './Drawer';
 
 export default {
     title: 'Molecules/Drawer',
     component: Drawer,
     decorators: [ withKnobs ],
     argTypes: {
-        
+
     },
 };
