+++ src/components/molecules/Drawer/Drawer.test.js
@@ -1,4 +1,4 @@
-import Drawer from '.';
+import Drawer from './Drawer';
 import runCommonTests from '../../../../test/commonTests';
 
 runCommonTests(Drawer);
