+++ src/components/molecules/Popup/Popup.test.js
@@ -1,4 +1,4 @@
-import Popup from '.';
+import Popup from './Popup';
 import runCommonTests from '../../../../test/commonTests';
 
 runCommonTests(Popup);
